bug fix - removed accumulated double to int error in xAxis bar and tick mark positioning

+++ xchart/src/main/java/com/xeiam/xchart/internal/chartpart/PlotContentBarChart.java
@@ -69,7 +69,7 @@ public class PlotContentBarChart extends PlotContent {
       }
     }
     int numBars = categories.size();
-    int gridStep = (int) (xTickSpace / (double) numBars);
+    double gridStep = xTickSpace / (double) numBars;
 
     // plot series
     int seriesCounter = 0;
